rate/review html
so this is pretty rough, but I wanted to push a form up for use. The star rating is showing up as squares instead of stars, but I will fix that. The CSS I threw in was just to make it a little easier on the eyes, but I plan on changing it up.